Add WalletStorage to easily access the wallet
It knows the wallet and the upay server and should
be used the the main class for interacting with the
wallet.

To remove a specific amount of value from the wallet
it will create a transaction which contains the
needed tokens.

Also transactions are used to add tokens to the wallet.